Fix Azure init but using bad time types

+++ Source/Node/nodes/Source/azure/KinectAzureNode.h
@@ -63,7 +63,7 @@ public:
 	Image colorImage;
 
 	bool newFrameAvailable;
-	long timeAtLastInit;
+	uint32 timeAtLastInit;
 
 	bool initInternal() override;
